+++ esp32/esp32_direct.ino
@@ -57,7 +57,7 @@ void handleGetReading() {
 
   // Field krusial untuk fitur AI
   doc["temperature"] = temperature;
-  doc["humidity"] = humidity;
+  doc["humidity"] = (int)humidity;
   doc["mq135_raw"] = mq135_raw;
   doc["mq135_ppm"] = mq135_ppm;
   doc["aqi"] = estimated_aqi;
